requirements.txt
Flask==2.3.2
Flask
gunicorn
psycopg2-binary
flask_sqlalchemy
